requirements.txt
alembic>=1.13.0
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic-settings==2.1.0
sqlalchemy==2.0.25
aiosqlite==0.19.0
python-dotenv==1.0.0
google-genai==0.2.2
openai
httpx
python-telegram-bot==20.7
